+++ extensions/duck-duck-go-search/src/utils/types.ts
@@ -1,0 +1,12 @@
+export const HISTORY_KEY = "history";
+
+export interface SearchResult {
+  id: string;
+  query: string;
+  url: string;
+  isHistory?: boolean;
+}
+
+export interface Preferences {
+  rememberSearchHistory: boolean;
+}