+++ extensions/quantumcast/src/open-cloudflow-manual.tsx
@@ -1,6 +1,7 @@
 import { open, getPreferenceValues } from '@raycast/api';
+import { cloudflow } from 'quantumlib';
 
 export default async () => {
   const { cloudflowBaseUrl } = getPreferenceValues();
-  await open(new URL(`${cloudflowBaseUrl}/manual/manual.html`).toString());
+  await open(cloudflow.getManualURL(cloudflowBaseUrl));
 };