+++ extensions/google-chrome/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Google Chrome Changelog
 
+## [Search url] - 2023-04-01
+
+- Added support to match url in bookmark and history search
+
+## [New Window] - 2023-03-22
+
+- Add the ability to create new window
+
 ## [Tab closing] - 2023-03-10
 
 - Add the ability to close tabs